+++ packages/fence_service/test/enrol_device_test.dart
@@ -1,11 +1,13 @@
 import 'package:fence_service/fence_service.dart';
 import 'package:fence_service/mongo_local_testing.dart';
 import 'package:protos_weebi/data_dummy.dart';
-import 'package:protos_weebi/grpc.dart';
+//import 'package:protos_weebi/grpc.dart';
 
 import 'package:test/test.dart';
 import 'package:mongo_dart/mongo_dart.dart';
 import 'package:protos_weebi/protos_weebi_io.dart';
+
+import 'service_call_impl.dart';
 
 void main() {
   final db = TestHelper.localDb;