+++ packages/protos/protos_weebi/lib/grpc.dart
@@ -1,2 +1,3 @@
 export 'package:grpc/grpc.dart';
-export 'package:protos_weebi/src/service_call_impl.dart';
+// ! do not export grpc_web because incompatible with app
+// export 'package:grpc/grpc_web.dart';